Refactor hash calculation
