+++ lib/hir-mir/src/lib.rs
@@ -31,7 +31,9 @@ pub use map_type_configuration::{
 };
 pub use string_type_configuration::StringTypeConfiguration;
 pub use test_module_configuration::TestModuleConfiguration;
-use transformation::{equal_operation, map_context, record_equal_function, record_hash_function};
+use transformation::{
+    equal_operation, hash_calculation, map_context, record_equal_function, record_hash_function,
+};
 
 pub fn compile_main(
     module: &Module,
@@ -79,6 +81,7 @@ fn compile_module(
     let module = record_hash_function::transform(&context, &module)?;
     let module = map_context::module::transform(&context, &module)?;
     let module = equal_operation::module::transform(&context, &module)?;
+    let module = hash_calculation::module::transform(&context, &module)?;
 
     Ok((
         {
